create cart.js module and modules folder
move store modules into modules folder
